improve filenames and __init__
