+++ elastodynamicsx/pde/materials/dampings.py
